% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{show,Gingr-method}
\alias{show,Gingr-method}
\title{Show method for Gingr objects}
\usage{
\S4method{show}{Gingr}(object)
}
\arguments{
\item{object}{an S4 \code{Gingr} class obect}
}
\description{
Show method for Gingr objects
}
