% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gingr.R
\name{gingr}
\alias{gingr}
\title{Material calculator to make gingerbread stuff...}
\usage{
gingr(area, units, structure = NULL, ...)
}
\arguments{
\item{area}{the numerical area you need}

\item{units}{the units you calculated the area in (in, ft, cm, m)}

\item{structure}{the type of gingerbread structure you wish to make}
}
\description{
This function will calculate the amount of material required to make a 
gingerbread related structure of any size!
}
\details{
Using a propietary 'secret' formula of baking goods values 
(Monier et al. 20xx), this function will precisely
calculate the necessary materials to make anything out of gingerbread. All
you need to enter is the unit area your mega structure will cover and it will
give you an output of what you need with respective units.
}
\examples{
gingr()
}
\author{
Brandon Monier, \email{brandon.monier@sdstate.edu}
}
