% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GingrEnglish-class}
\alias{GingrEnglish-class}
\alias{InEnglishPls}
\title{Class GingrEnglish}
\usage{
InEnglishPls(object)
}
\arguments{
\item{object}{an \code{Gingr} class object}
}
\description{
Class \code{GingrEnglish} defines a Gingr data set for
   constructing gingerbread-related things in English units.

This function will convert the display to English units for
   \code{Gingr} classes.
}
\examples{
gDS <- Gingr(area = 500, unit = "in", structure = "House")
InEnglishPls(gDS)

}
\author{
Brandon Monier, \email{brandon.monier@gmail.com}
}
