% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{show,GingrEnglish-method}
\alias{show,GingrEnglish-method}
\title{Show method for GingrEnglish objects}
\usage{
\S4method{show}{GingrEnglish}(object)
}
\arguments{
\item{object}{an S4 \code{GingrEnglish} class object}
}
\description{
Show method for GingrEnglish objects
}
