% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{gingrFact}
\alias{gingrFact}
\alias{gingrFact,Gingr-method}
\title{Get facts from a Gingr object}
\usage{
gingrFact(object, character)

\S4method{gingrFact}{Gingr}(object, character = "elf")
}
\arguments{
\item{object}{an S4 \code{Gingr} class object}

\item{character}{what character do you want to say the fortune? Current
options are \code{"elf", "devil", or "alien"}.}
}
\description{
Get facts from a Gingr object

Get a random fortune from Gingr objects
}
