% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{recipe}
\alias{recipe}
\alias{recipe,Gingr-method}
\title{Get recipe from a Gingr object}
\usage{
recipe(object)

\S4method{recipe}{Gingr}(object)
}
\arguments{
\item{object}{an S4 \code{Gingr} class object}
}
\description{
Get recipe from a Gingr object

Getter method for recipe extraction of Gingr objects
}
