% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\docType{methods}
\name{gingrFact}
\alias{gingrFact}
\alias{gingrFact,Gingr-method}
\title{Get facts from a Gingr object}
\usage{
gingrFact(object)

\S4method{gingrFact}{Gingr}(object)
}
\arguments{
\item{object}{an S4 \code{Gingr} class object}
}
\description{
Get facts from a Gingr object

Get a random fortune from Gingr objects
}
